Fixed 2801807 - DateTime class is no longer ZORBA_DLL_PUBLIC. Moved util/time.h from src/ to include/. It now provides cross-platform timing utilities for CPU and wall-clock timing, and uses POSIX high-precision timers where available. zorbacmd and fn:doc() use this rather than DateTime for benchmarking.

+++ include/zorba/xquery.h
@@ -515,7 +515,7 @@ class ZORBA_DLL_PUBLIC XQuery : public SmartObject
   /**
    *
    */
-  virtual long
+  virtual double
   getDocLoadingTime() const = 0;
 };
